auth-service/requirements.txt
# NetBet Auth Service - Betfair API Authentication
requests>=2.31.0
python-dotenv>=1.0.0
cryptography>=42.0.0
